+++ web-app/src/css/Auth.css
@@ -1,0 +1,6 @@
+.auth {
+	height: 100%;
+	width: 100%;
+	background: url("../assets/images/banner.png");
+	background-size: cover;
+}
