+++ web-app/src/css/ForgotPassword.css
@@ -1,0 +1,23 @@
+.forgot-password-section {
+	position: fixed;
+	top: 50%;
+	left: 50%;
+	-webkit-transform: translate(-50%, -50%);
+	transform: translate(-50%, -50%);
+	width: 1000px;
+	/* height: 500px; */
+	/* margin: 0 auto; */
+	box-shadow: 0px 0px 10px 5px rgba(0, 0, 0, 0.1);
+}
+
+.forgot-password {
+	position: fixed;
+	top: 50%;
+	left: 50%;
+	-webkit-transform: translate(-50%, -50%);
+	transform: translate(-50%, -50%);
+}
+
+.fields-pass {
+	width: 400px;
+}
