+++ web-app/src/css/Login.css
@@ -1,7 +1,12 @@
 .login-section {
+	position: fixed;
+	top: 50%;
+	left: 50%;
+	-webkit-transform: translate(-50%, -50%);
+	transform: translate(-50%, -50%);
 	width: 1000px;
 	height: 500px;
-	margin: 0 auto;
+	/* margin: 0 auto; */
 	box-shadow: 0px 0px 10px 5px rgba(0, 0, 0, 0.1);
 }
