+++ web-app/src/css/Register.css
@@ -1,8 +1,13 @@
 .register-section {
+	position: fixed;
+	top: 50%;
+	left: 50%;
+	-webkit-transform: translate(-50%, -50%);
+	transform: translate(-50%, -50%);
 	width: 1000px;
-	height: 685px;
+	/* height: 685px; */
 	box-shadow: 0px 0px 10px 5px rgba(0, 0, 0, 0.1);
-	margin: 0 auto;
+	/* margin: 0 auto; */
 }
 
 .upper-part {
@@ -39,7 +44,7 @@
 .form-section {
 	padding-top: 2%;
 	background: #ffffff;
-	height: 78%;
+	height: 100%;
 }
 
 .gold-btn {
